working threaded for file 1
