+++ etl/shared_funcs.py
@@ -1,415 +1,242 @@
-# Functions common to all four of the fact table processing scripts.
-# Most of the functionality is in here or at least goes through here.
-# This could likely be reused as a library for other ETL tasks by updating table_info.py and dimension.py to match different ETL tasks
+from __future__ import annotations
 
 
-# import python libraries
-import time
-import MySQLdb
-import clickhouse_driver
-from collections import deque
+# TODO get rid of this global variable
+line_num = 0
 
-# import my python modules
-import dimension
-import table_info
-
-
-# dimension data structures like {name: dimension_object}
-dicts = {} #TODO rename this
-
-# todo probably better saving these to a file directly. 
-# also adding a max error limit to avoid going through the whole dataset doing nothing
-# for error handling, unused.
-error_rows = []
-# for expectedly skipped rows. currently also unused.
-skipped_rows = []
-district_rows = []
-
-# TODO consider renaming places where dim_dict is to just dimension and similar (this file's  object dicts)
-
-# mostly for backwards compatibility purposes atm
-def _new_dim_dict(name):
-    """
-    Create a dimension object and add it to this module's dimension dictionary under the name provided.
-    If name is 'time', creates TimeDimension, if 'location' LocationDimension, otherwise SimpleDimension.
-
-    Raises ValueError is a dimension with this name already exists in the dictionary.
-    """
-
-    # todo consider checking against table_info.ALLOWED_TABLES before creating
-
-    if name in dicts:
-        raise ValueError(f'Dict with this name already exists: {name}')
-    if name == 'time':
-        dicts[name] = dimension.TimeDimension()
-    elif name == 'location':
-        dicts[name] = dimension.LocationDimension()
-    else:
-        dicts[name] = dimension.SimpleDimension()
-
-
-# consider combining with _new_dim_dict into one
-def _ensure_dim_dict(name):
-    """Calls _new_dim_dict if a dimension with this name does not exist already."""
-    if name not in dicts:
-        _new_dim_dict(name)
-
-# consider stopping supporting this functionality, it's not even that important
-def _unwrap_wrap_args(*args):
-    """Turns arguments like ('a', ['b', 'c'], 'd') into list ['a', 'b', 'c', 'd']."""
-    out = []
-    for arg in args:
-        if type(arg) == list or type(arg) == tuple:
-            for element in arg:
-                out.append(element)
-        else:
-            out.append(arg)
-    return out
-
-def ensure_dim_dicts(*dimension_names):
-    """
-    Creates dimension objects if ones with this name do not exist yet.
-    Expected names of dimensions are in table_info.ALLOWED_TABLES.
-    """
-    
-    # in retrospect, accepting either a list or a *args instead of just one of them is a weird way to do things.
-    dims_list = _unwrap_wrap_args(dimension_names)
-    for dim in dims_list:
-        _ensure_dim_dict(dim)
-
-
-def split_line(line):
-    """Remove trailing whitespace and split on every comma"""
-    return line.rstrip().split('|')
-
-
-def create_db_table(table_name, dbcon, drop_if_exists = True):
-    """Create a table in the database using table_info.create_string[table_name]."""
-    c = dbcon.cursor()
-    if (drop_if_exists):
-        # maybe be more carfeul with dropping in production
-        # also maybe escape
-        c.execute(f'DROP TABLE IF EXISTS {table_name}')
-        # c.execute('DROP TABLE IF EXISTS %s', table_name)
-    c.execute(table_info.create_strings[table_name])
-
-
-def __raise_not_numeric(row_cell, row):
-    raise ValueError('All elements in the fact table must be numerical to prevent SQL injection attacks.' +
-        f'Instead found: element {row_cell} of type {type(row_cell)} in row {row}')
-
-def _row_cell_to_numbers(row_cell, row = None):
-
-    if(type(row_cell) == int or type(row_cell) == float):
-        return row_cell
-    else:
-        # raise an exception if not a number but chceck numeric strings
-        if (type(row_cell) == str):
-            try:
-                # throws a ValueError if string not convertible to int
-                row_cell = int(row_cell)
-                return row_cell
-            except ValueError:
-                try:
-                    # throw a ValueError if string not convertible to float
-                    row_cell = float(row_cell)
-                    return row_cell
-                except ValueError:
-                    __raise_not_numeric(row_cell, row)
-        else:
-            __raise_not_numeric(row_cell, row)
-
-def _raise_if_not_int_float(list_of_rows):
-    # out = []
-    for i,row in enumerate(list_of_rows):
-        list_of_rows[i] = [_row_cell_to_numbers(row_cell) for row_cell in row]
-        # print(row)
-        # print(list_of_rows[i])
-        # out.append(row)
-    # return out
-
-def save_batch(batch, dbcon, table_name):
-    """
-    Save a batch of fact table rows into the database.
-
-    CAREFUL: MAY be susceptible to SQL injection attacks. 
-        Does not use existing library functions to escape any characters but only accepts int or float.
-        So should be safe but do procees with caution.
-    """
-
-    values = list(batch)
-
-    # prepare the headers part of the SQL query
-    headers = table_info.headers_dict[table_name]
-    headers_str = '('+ ','.join(headers) + ')'
-
-    sql_base = f"INSERT INTO {table_name} {headers_str} VALUES "
-
-    # raise an Exception if anything of type other than (float, int) comes in
-    # be very careful removing the following line, will need to escape string if anything other than numbers comes in.
-    # print(1)
-    # values = 
-    _raise_if_not_int_float(values)
-    # print(2)
-    # print(values[:10])
-    dbcon.cursor().execute(sql_base, values)
-    # print(f'saved to ch')
-
-    # each row in values is a tuple so str(v) is (e1, e2, e3, ...) already, and just put commas in between each
-    # values_str = ','.join([str(v) for v in values])
-    # bigsql = sql_base + values_str
-    # if values_str == "":
-    #     raise Exception("batch was empty in save_batch")
-    # 
-    # dbcon.cursor().execute(bigsql)
-    # dbcon.commit()
-
-def save_batch_timed(*args):
-    """Call save_batch and time how long it took"""
-    start_time = time.time()
-    save_batch(*args)
-    end_time = time.time()
-    return end_time - start_time
-
-
-#TODO comments and better names
-
-def _read_filestream_as_split_lines(file):
-    for line in file:
-        yield split_line(line)
-
-def read_file_as_split_lines(filepath, skip_headers = True):
-    with open(filepath) as file:
-        if skip_headers:
-            _ = next(file)
-        
-        yield from _read_filestream_as_split_lines(file)
-
-def batch_process_threaded_from_file(path, skip_headers=True, **kwargs):
-    batch_process_threaded_from_generator(read_file_as_split_lines(path, skip_headers), **kwargs)
-
-def does_table_exist(table_name, dbcon):
-    c = dbcon.cursor()
-    c.execute("show tables like %s", (table_name,))
-    res = c.fetchone()
-    res = res[0]
-    return True if res == table_name else False
-
-def ensure_table(table_name, dbcon, if_exists):
-    if if_exists == 'replace':
-        return create_db_table(table_name, dbcon)
-    elif if_exists == 'append':
-        if not does_table_exist(table_name, dbcon):
-            # only create if it doesn't exist
-            create_db_table(table_name, dbcon)
-    elif if_exists == 'fail':
-        if does_table_exist(table_name, dbcon):
-            raise ValueError(f'Table {table_name} already exists')
-
-def save_errors(error_rows):
-    with open('error_file.txt', 'a') as file:
-        # file.write('\n')
-        file.write(str([f'{time.time()}: {str(e)}\n' for e in error_rows]))
-
-def clear_error_file():
-    with open('error_file.txt', 'w') as file:
-        file.write('')
-
-import concurrent.futures
-def batch_process_threaded_from_generator(row_generator, row_func, fact_table_name, dbcon, if_exists='fail', **row_func_kwargs):
-    start_time = time.time()
-    maxlen = 4000 # 16s. 1000 21s, 10000 17s
-    error_rows= []
-    clear_error_file()
-    batch = deque([], maxlen=maxlen)
-    # also consider multiprocessing Queue - one shared queue between threads, and just save as it fills up (or like a queue of batches even)
-
-    # connect if dbcon is null
-    connected_here = False
-    if not dbcon:
-        dbcon = connect_to_db()
-        connected_here = True
-
-    ensure_table(fact_table_name, dbcon, if_exists)
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        latest_future = None
-        total_insert_time = 0
-
-        for row in row_generator:
-            # also option of using csv.reader - slower but handles quotation marks around cells
-
-            # maybe make fact_line come from a generator instead, to separate creating and saving
-            try:
-                fact_line = row_func(row, **row_func_kwargs)
-                # except Exception as e:
-                #     error_rows.append((row, e))
-
-                batch.append(fact_line)
-            except dimension.DistrictException as de:
-                district_rows.append(row)
-                continue # skip the rest of the execution for this line here
-            except KeyboardInterrupt as ke:
-                save_errors(error_rows[:100])
-                raise ke
-            except Exception as e:
-                error_rows.append(tuple(e, row))
-                if (len(error_rows)>100):
-                    # too many errors
-                    raise e
-
-            if len(batch) == maxlen:
-                
-                try:
-                    if latest_future:
-                        # wait up to 2 seconds for the last SQL INSERT to finish (or raise an Exception)
-                        time_taken = latest_future.result(3)
-                        total_insert_time += time_taken
-
-                    latest_future = executor.submit(save_batch_timed, batch, dbcon, fact_table_name)
-                    batch = deque([], maxlen=maxlen) # empty the queue here
-                # except MySQLdb.DataError as mde:
-                #     error_rows.append((mde, batch))
-                # except TimeoutError
-                except KeyboardInterrupt as ke:
-                    raise ke
-                except Exception as e:
-                    error_rows.append((e, batch))
-                    if (len(error_rows)>100):
-                        # too many errors
-                        raise e
-            
-            if error_rows and len(error_rows) >=1000:
-                save_errors(error_rows)
-                error_rows = []
-        try:
-                
-            # wait for last save task to finish
-            if latest_future:
-                time_taken = latest_future.result(2)
-                total_insert_time += time_taken
-
-            # save items left in the queue after finished reading the file (e.g. if there are 9500 rows and batches are size 1000, 500 left at the end)
-            if len(batch) != 0:
-                latest_future = executor.submit(save_batch_timed, batch, dbcon, fact_table_name)
-                time_taken = latest_future.result(2)
-                total_insert_time += time_taken
-        except Exception as e:
-            error_rows.append(e, batch)
-
-        if error_rows:
-            save_errors(error_rows)
-
-        # close connection if connected here
-        if (connected_here):
-            dbcon.close()
-    end_time = time.time()
-    print(f'read handle and write took {end_time - start_time} seconds to table {fact_table_name}.')
-    print(f'Inserting into fact table took a total of {total_insert_time} seconds')
-    pass  
-
-
-# TODO rename this
-# TODO refactor needed, there is a lot of deeply nested function calls where each don't really add anything 
-def batch_process(path, row_func, fact_table_name, dbcon=None, **kwargs):
-    """Process file at path. See batch_process_threaded"""
-    batch_process_threaded_from_file(path=path, skip_headers=True, row_func=row_func, fact_table_name=fact_table_name, dbcon=dbcon, if_exists='replace', **kwargs)
-    pass
-
-
-def save_dim(name, dbcon = None):
-    """Wrapper for Dimension.to_sql, also creates connection if dbcon null"""
-    db_connected_here = False
-    if not dbcon:
-        dbcon = connect_to_db()
-        db_connected_here = True
-
-    dicts[name].to_sql(name, dbcon)
-    if db_connected_here:
-        dbcon.commit()
-        dbcon.close()
-
-def save_dims():
-    """Save all the created dimensions into SQL database"""
-    dbcon = connect_to_db()
-    for name in dicts.keys():
-        save_dim(name, dbcon)
-    dbcon.commit()
-    dbcon.close()
-
-class ClickhouseCursorClient:
-    def __init__(self, client):
-        self.client = client
-    def cursor(self):
-        return self
-    def commit(self):
-        pass
-    def close(self):
-        self.client.disconnect()
-    def execute(self, *args, **kwargs):
-        return self.client.execute(*args, **kwargs)
-    def executemany(self, sql, values):
-        sqlnoval = sql.split('%')[0][:-1]
-        return self.client.execute(sqlnoval, values)
-        print(sql)
-        print(values)
-
-dbpassword = None
+dbpassword = 'RadicalSpiderWearingPaper'
 
 def getDbPassword():
+    global dbpassword
     if dbpassword:
         return dbpassword
     else:
         dbpassword = input("Please enter the database password ")
         return dbpassword
-# TODO make this adjustable
-def connect_to_db():# -> MySQLdb.Connection:
-    """Get MySQLdb connection to default database."""
-    db = clickhouse_driver.Client(host='localhost', password = dbpassword, database='sgov')
-    db = ClickhouseCursorClient(db)
-    # db = MySQLdb.connect(host='localhost:9004',user='temp_user', passwd = 'password', database='sgov')
-    # import psycopg2
-    # db = psycopg2.connect(database='sgov_mini', user='temp_user', password='password')
-    # c = db.cursor()
-    return db
 
-def get_sqlalchemy_con():
-    """Get SQLAlchemy engine for the database"""
-    import sqlalchemy
-    return sqlalchemy.create_engine('clickhouse://', creator= lambda :connect_to_db().client)
-    # return sqlalchemy.create_engine('mysql://', creator=connect_to_db)
 
-def get_sql_alchemy_con_for_read():
-    import sqlalchemy
-    return sqlalchemy.create_engine(f'clickhouse+native://default:{dbpassword}@localhost:8123')
-    return sqlalchemy.create_engine('clickhouse://', creator= lambda :connect_to_db())
+#TODO load from a file
+def get_zip_pwd(infpath):
+    import os
+    return {
+        'module1_sample_10k.zip': None,
+        'nr_module_1_2020.csv.zip': b'NetworkRail_2020',
+        'nr_module_1_2021.csv.zip': b'NetworkRail_2021',
+        'nr_module_1_2022.csv.zip': b'NetworkRail_2022',
+    }[os.path.basename(infpath)]
 
-# A lot of wrapper functions which are mostly here for historic purposes. But may be good to keep anyway in case changes are wanted.
-#TODO change explicit args to *args or **kwargs
+def get_filename_in_zip(infpath):
+    import os
+    return {
+        'module1_sample_10k.zip': 'module1_sample_10k.csv',
+        'nr_module_1_2020.csv.zip': 'san-ssapfs/edge/home/chaudhup/Network_Rail/nr_module_1_2020.csv',
+        'nr_module_1_2021.csv.zip': 'san-ssapfs/edge/home/chaudhup/Network_Rail/nr_module_1_2021.csv',
+        'nr_module_1_2022.csv.zip': 'san-ssapfs/edge/home/chaudhup/Network_Rail/nr_module_1_2022.csv',
+    }[os.path.basename(infpath)]
 
-def add_to_dict_if_not_in(dict_name, key, values: list):
-    """Wrapper function for dimension.add_if_not_in. Takes in name of dim as stored in dicts here"""
-    return dicts[dict_name].add_if_not_in(key, values)
+def split_row_generator(infpaths, colsep, print_headers = True):
+    for infpath in infpaths:
+        import zipfile
+        with zipfile.ZipFile(infpath) as zf:
+            with zf.open(get_filename_in_zip(infpath), pwd = get_zip_pwd(infpath)) as infile:
+                headers = next(infile)
+                if print_headers:
+                    print(headers)
 
- 
-def handle_one_dim(dict_name, row, row_index):
-    """Wrapper for SimpleDimension.add_row. Takes in name of dim as stored in dicts here"""
-    return dicts[dict_name].add_row(row, row_index)
+                for line in infile:
+                    global line_num
+                    line_num += 1
 
-def handle_time(row, index = 0):
-    """Wrapper for TimeDimension.add_row. Takes in name of dim as stored in dicts here"""
-    return dicts['time'].add_row(row, index)
+                    # if isinstance(line, bytes):
+                    line = line.decode('utf-8')
+                    linestrip = line.rstrip()
+                    split = linestrip.split(colsep)
+                    yield split
 
-def handle_loc(row, loc_level_idx = 4, loc_idx = 5, smallest_area_name = 'POSTCODE_SECTOR', skip_bigger_area_rows=False):
-    """Wrapper for LocationDimension.add_row. Takes in name of dim as stored in dicts here"""
-    return dicts['location'].add_row(row, loc_level_idx, loc_idx, smallest_area_name, skip_bigger_area_rows=skip_bigger_area_rows)
 
-# consider removing
-def postcode_sector_to_loc_list(sector: str):
-    """Wrapper for LocationDimension.postcode_sector_to_loc_list."""
-    return dimension.LocationDimension.postcode_sector_to_loc_list(sector)
+import concurrent.futures
+class CsvDatabaseWriterThreaded:
+    def __init__(self, outfpath, batch_size, dbname, table_name, timeout_s = 10) -> None:
+        self.executor = None
+        self.latest_future = None
+        self.swapfpath = outfpath + '_2'
+        self.timeout_s = timeout_s
+        # super().__init__(outfpath, batch_size, dbname, table_name)
 
-def get_headers(dict_name):
-    """Get the headers of a given dimension"""
-    return dicts[dict_name].headers
+        self.dbname = dbname
+        self.batch_size = batch_size
+        self.outfpath = outfpath
+        self.table_name = table_name
+        self.line_counter = 0
 
+    def __enter__(self):
+        self.executor = concurrent.futures.ThreadPoolExecutor()
+        # return super().__enter__()
+        self.open()
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+
+        # first check if any lines left in current batch
+        if self.line_counter % self.batch_size != 0:
+            # self.close_file_for_saving()
+            self._save_to_db_single(self.outfile, self.outfpath)
+        self.outfile.__exit__(exc_type, exc_value, traceback)
+        self.executor.__exit__(exc_type, exc_value, traceback)
+        # return super().__exit__(exc_type, exc_value, traceback)
+
+    def open(self, ):
+        self.outfile = open(self.outfpath, 'w')
+        return self.outfile
+
+    def close(self):
+        return self.__exit__(None, None, None)
+
+    def _save_to_db_single(self, file_context, file_name, num_header_lines = 0):
+        # need to take in the file_context and file_name as parameters to avoid race conditions
+        file_context.close()
+
+        import subprocess
+        subprocess.Popen((f'clickhouse-client --password={getDbPassword()}'
+        ' --format_csv_delimiter="|"'
+        f' --input_format_csv_skip_first_lines={num_header_lines}'
+        f' --query="INSERT INTO {self.dbname}.{self.table_name} FORMAT CSV" < "{file_name}"'),
+        shell=True).wait()
+
+
+
+    def re_open(self):
+        self.outfile = open(self.outfpath, 'w')
+
+    def _save_to_db_threaded(self, *args, **kwargs):
+        # wait until last thread finished
+        if self.latest_future is not None:
+            self.latest_future.result(self.timeout_s)
+
+        self.latest_future = self.executor.submit(self._save_to_db_single,
+            self.outfile, self.outfpath, *args, **kwargs)
+        self.outfpath, self.swapfpath = self.swapfpath, self.outfpath
+        # re_open happens in write()
+
+    def write(self, line, colsep):
+        return_value = self.outfile.write(colsep.join(line)+ '\n')
+        # save to csv if batch filled
+        self.line_counter += 1
+        if self.line_counter % self.batch_size == 0:
+            # self.close_file_for_saving()
+            self._save_to_db_threaded()
+            self.re_open()
+        return return_value
+
+
+class ErrorHandler:
+    def __init__(self, errorfpath, max_errors_before_crash = None, colsep = '|') -> None:
+        self.errorfpath = errorfpath
+        self.max_errors_before_crash = max_errors_before_crash
+        self.colsep = colsep
+        self.error_count = 0
+        pass
+    
+    def __enter__(self):
+        self.open()
+        return self
+    def __exit__(self, type = None, value = None, traceback = None):
+        self.errorfile.__exit__(type, value, traceback)
+        
+
+    def open(self):
+        self.errorfile = open(self.errorfpath, 'w')
+        return self.errorfile
+
+    def close(self, *args, **kwargs):
+        return self.__exit__()
+
+
+    def line_arr_to_str(self, line_arr):
+        return self.colsep.join(line_arr)
+
+    def save_and_ignore_until_max(self, line):
+        pass
+        if isinstance(line, list):
+            line = self.line_arr_to_str(line)
+        print(line)
+        import traceback
+        tb = traceback.format_exc()
+        print(tb)
+        self.errorfile.write(line)
+        self.errorfile.write('#')
+        self.errorfile.write(str(tb))
+        self.errorfile.write('\n')
+        self.error_count += 1
+        if self.max_errors_before_crash is not None:
+            if self.error_count > self.max_errors_before_crash:
+                raise 
+
+
+
+class DBInfo:
+    """
+    Be VERY careful with this class, it may lead to SQL injection attacks 
+    if any user-provided data is used here. 
+    (Limiting types to string literals will become available in Python 3.11)
+    """
+    def __init__(self, creation_string_columns, dbname = None, table_name = None) -> None:
+        self.dbname = dbname
+        self.table_name = table_name
+        self.creation_string_columns = creation_string_columns
+
+    def _ensure_fields_not_None(self):
+        if self.creation_string_columns is None:
+            raise RuntimeError('DBInfo creation_string_columns is not set.')
+        if self.dbname is None:
+            raise RuntimeError('DBInfo dbname is not set.')
+        if self.table_name is None:
+            raise RuntimeError('DBInfo table_name is not set.')
+
+
+    def csv_to_db_create(self, drop_if_exists = True):
+        
+        self._ensure_fields_not_None()
+        import os
+
+        if drop_if_exists:
+            os.system(f'clickhouse-client --password={getDbPassword()}'
+            f' --query="DROP TABLE IF EXISTS {self.dbname}.{self.table_name};"')
+
+        os.system(f'clickhouse-client --password={getDbPassword()}'
+            ' --query="'
+            f"CREATE TABLE {self.dbname}.{self.table_name} ("
+            f'{self.creation_string_columns}'
+            ';"')
+
+def apply_and_save(infpaths, row_function, dbinfo, *extra_params, **extra_named_parameters):
+
+    # todo look into alternative file types for the intermediate file, csv too big (zipped may be fine)
+    # outfpath = '/mnt/sftp/module 1/tmp_module1_t1.csv'
+    outfpath = 'data/tmp_to_save.csv'
+    errorfpath = 'error_file.txt'
+    max_errors_before_crash = 1#00000
+    if dbinfo.dbname is None:
+        dbinfo.dbname = 'nr'
+    # consider moving this to CsvDatabaseWriter, with an optional append vs replace parameter
+    dbinfo.csv_to_db_create()
+    colsep = '|'
+
+
+    with CsvDatabaseWriterThreaded(outfpath, 1002, dbinfo.dbname, dbinfo.table_name) as writer, \
+            ErrorHandler(errorfpath, max_errors_before_crash, colsep) as error_handler:
+        for split in split_row_generator(infpaths, colsep):
+            try:
+
+                conv = row_function(split, *extra_params, *extra_named_parameters)
+                
+                writer.write(conv, colsep)
+
+            except KeyboardInterrupt as ki:
+                print(f'KeyboardInterrupt at infile line {line_num}')
+                raise ki
+            except Exception as e:
+                error_handler.save_and_ignore_until_max(split)
+                pass
+        print(f'Loaded {writer.line_counter} lines into the database.')
